Address latest review round: flatten Python reference, drop Rust reference, specify Tweak Server
- Flatten python/elip_sp_reference/{__init__,core,bech32m}.py into a single
  reference.py + bech32m.py alongside the ELIP, matching bitcoin/bips'
  bip-0352/ flat-script layout. Drop the package's pyproject.toml, README.md,
  .gitignore and __init__.py re-export shim per reviewer request.
- Trim verbose AI-generated-sounding docstrings/comments to what's needed.
- Remove the Rust reference implementation entirely (rust/) so there is one
  reference implementation, in the BIP-352 reference.py style, as requested.
- Add a normative Tweak Server subsection to the ELIP: the server publishes
  T = input_hash*A per transaction and derives S = b_scan*T; previously only
  implemented in code (compute_tweak/shared_secret_from_tweak) but not
  specified in the document.
- Update Reference Implementation section and README to match the new layout.
