+++ elip-silent-payments-liquid/README.md
@@ -1,17 +1,21 @@
-# elip-silent-payments-liquid — reference implementations
+# elip-silent-payments-liquid — reference implementation
 
-Reference implementations for [`elip-silent-payments-liquid.mediawiki`](../elip-silent-payments-liquid.mediawiki),
-following the layout convention of [bitcoin/bips](https://github.com/bitcoin/bips)
-(e.g. `bip-0352/`): supporting code lives in a folder named after the ELIP.
+Reference implementation for [`elip-silent-payments-liquid.mediawiki`](../elip-silent-payments-liquid.mediawiki),
+following the layout of [bitcoin/bips](https://github.com/bitcoin/bips) `bip-0352/`:
+a flat script next to the ELIP, in the style of the canonical
+[BIP-352 reference](https://github.com/bitcoin/bips/blob/master/bip-0352/reference.py).
 
-Two independent implementations derive the same test vectors byte-for-byte:
+- [`reference.py`](reference.py) — the protocol: addresses, input aggregation,
+  shared secret, per-output derivation, tweak server, and Confidential
+  Transactions blinding/unblinding. Elliptic-curve algebra uses the pure-Python
+  `secp256k1lab` (same library the BIP-352 reference uses); the CT-specific
+  functions use `wallycore`.
+- [`bech32m.py`](bech32m.py) — Bech32m, verbatim from BIP-350 / the BIP-352
+  reference.
+- [`test_vectors.py`](test_vectors.py) — byte-pinned test vectors plus
+  taproot/tweak-server/address/CT round-trip tests.
 
-- [`rust/`](rust) — built on `lwk_wollet`. `cargo test` to run.
-- [`python/`](python) — in the canonical BIP-352 `reference.py` style, using the
-  pure-Python `secp256k1lab` for the curve algebra and `wallycore` only for the
-  Liquid Confidential Transactions plumbing. See [`python/README.md`](python/README.md).
-
-Both cover:
+Covers:
 
 - **Address encoding** — Bech32m, HRP `lqsp`/`tlqsp`, version `q`
 - **Sender derivation** — input aggregation, ECDH shared secret, `P_k`, `BK_k`, `bk_k`
@@ -19,7 +23,17 @@ Both cover:
 - **Receiver scanning** — recompute `P_k`, match against outputs, derive spend secret
 - **Confidential output blinding and unblinding** — the ELIP's novel claim: `bk_k`
   derived from the shared secret unblinds the output non-interactively
-- **Test vectors** — deterministic known-answer values matching the ELIP specification
+
+## Running
+
+```sh
+pip install pytest 'secp256k1lab @ git+https://github.com/secp256k1lab/secp256k1lab.git' wallycore
+pytest -v elip-silent-payments-liquid/test_vectors.py
+```
+
+`secp256k1lab` is fetched from git (not published on PyPI). `wallycore`'s PyPI
+wheels are built with Elements support, which the CT test needs; if it is not
+installed, that one test is skipped automatically.
 
 ## License
