+++ elip-silent-payments-liquid/bech32m.py
