--- elip-silent-payments-liquid/python/.gitignore
@@ -1,13 +1,0 @@
-# Virtual environments
-.venv/
-venv/
-
-# Build / packaging artifacts
-*.egg-info/
-build/
-dist/
-
-# Caches
-__pycache__/
-*.py[cod]
-.pytest_cache/
